Implement meta-ladders 9 and 0
This assumes that the SRPR and AF values for ladder 9
is a roll-up of ladders 4, 5, and 8 with value = MAX(value).
Meaning it's the best of the GX ladders.

Ladder 0 is a roll-up of the ladders 1, 2, 3, 6, 7, 9 with value =
SUM(value). Meaning it's the sum of every game.

Not sure what the ladders 11-18 represent yet.