Remove b, europeanunion, jacorre, and united nations flags.
b.gif is either a placeholder or layout image which isn't used in the current code.
jacorre seems to be the name of the site that the 27x17 flags were from.
europeanunion and united nations are the only 'super-national' flags here. They probably don't fit in and seems unlikely anyone will use them.