Write game rules in Markdown instead of HTML, making it easier to edit.
michelf/php-markdown is used to convert the Markdown into HTML to
display on the rules pages.

Note how numbered lists use "1." for every item. The Markdown renderer
will fill in the actual increasing numbers automatically, and leaving
the actual number unspecified in the Markdown means that diffs will be
more concise when items are added/removed.