Move the "FZC reserves the right to request proof..." rule to all.md.
Climax / GPL / MV / X had a version that said photo proof was
acceptable, but this probably hasn't been the case for a long time, so
the GX version that doesn't reference photo proof is now used.

SNES just didn't have this rule, but don't see why it wouldn't apply the
same. (Alternatively, this rule could just be self-evident and not
needed; but for now, it's still in.)